fix bug on profiler class detection, fix bug on old toolbar cleaning

+++ Helper/Data.php
@@ -319,7 +319,9 @@ class Data extends AbstractHelper
 
             $folder = $this->getToolbarFolder();
             foreach ($toDelete as $file) {
-                unlink($folder . $file);
+                if (is_file($folder . $file)) {
+                    unlink($folder . $file);
+                }
             }
         }
     }
